+++ thoth-api/migrations/v0.11.19/down.sql
@@ -1,0 +1,2 @@
+ALTER TABLE imprint
+    DROP COLUMN crossmark_doi;