+++ thoth-app/src/models/imprint/update_imprint_mutation.rs
@@ -1,6 +1,7 @@
 use serde::Deserialize;
 use serde::Serialize;
 use thoth_api::model::imprint::Imprint;
+use thoth_api::model::Doi;
 use uuid::Uuid;
 
 const UPDATE_IMPRINT_MUTATION: &str = "
@@ -8,12 +9,14 @@ const UPDATE_IMPRINT_MUTATION: &str = "
         $imprintId: Uuid!,
         $imprintName: String!,
         $imprintUrl: String,
+        $crossmarkDoi: Doi,
         $publisherId: Uuid!
     ) {
         updateImprint(data: {
             imprintId: $imprintId
             imprintName: $imprintName
             imprintUrl: $imprintUrl
+            crossmarkDoi: $crossmarkDoi
             publisherId: $publisherId
         }){
             imprintId
@@ -42,6 +45,7 @@ pub struct Variables {
     pub imprint_id: Uuid,
     pub imprint_name: String,
     pub imprint_url: Option<String>,
+    pub crossmark_doi: Option<Doi>,
     pub publisher_id: Uuid,
 }